Move entity management to separate controllers
